changed pin toggle to work on a true / false depending on the pin toggle position - 100% working, and will always give accurate results.
Changed CustomAdapter to work with the flagAndCurrency list size, instead of resting/using 32 (full country list size).  This is done in prep for the SettingsActivity / SettingsAdapter where mast currencies will be removed.

work needs to be done in MainActivity on the populatedArrayList which returns ArrayList  - flagAndCurrencyList at bottom.  possibly remove positions from positionsToRemove first in a separate try/catch (from settingsActivity) then go on to remove additional positions IF the menu toggle is set on for the pinned positions.